PriceAPI/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
idna==3.4
pytz==2023.3
requests==2.29.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
